# 03_openapi_original.yaml
openapi: 3.1.0
info:
  title: Kaveri Stays API
  version: 1.0.0
  description: >
    Production-grade HTTP API for Kaveri Stays luxury boutique resort chain.
    Enforces role-based permissions (guest, staff, manager, owner), transactional guarantees,
    and PostgreSQL exclusion-backed booking boundaries.
servers:
  - url: http://127.0.0.1:8000
    description: Local Development Server
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new guest account
      description: Creates a self-service guest account. Cannot assign staff/manager/owner roles.
      operationId: register_account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login and obtain JWT access & refresh tokens
      description: Authenticates user credentials with bcrypt/Argon2. Rate-limited to prevent brute-force attacks.
      operationId: login_account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Rotates the refresh token. Invalidates old refresh token to prevent replay.
      operationId: refresh_token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout and revoke refresh token
      operationId: logout_account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      tags:
        - Authentication
      summary: Get current authenticated user profile
      operationId: get_current_user_profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: "Aarav Sharma"
        email:
          type: string
          format: email
          example: "aarav@example.com"
        password:
          type: string
          minLength: 8
          example: "StrongPassword123!"
        phone:
          type: string
          example: "+91 98765 43210"
        city:
          type: string
          example: "Bengaluru"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "guest@example.com"
        password:
          type: string
          example: "Password123!"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 900
    AccountResponse:
      type: object
      required: [account_id, name, email, role]
      properties:
        account_id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [guest, staff, manager, owner]
        property_id:
          type: integer
          nullable: true
        guest_id:
          type: integer
          nullable: true
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "ROOM_UNAVAILABLE"
            message:
              type: string
              example: "The room is not available for the requested dates."
            details:
              type: array
              items:
                type: object